feat: refactor plugin tree view compoent
