+++ owlplug-client/src/main/resources/fxml/PluginsView.fxml
@@ -14,13 +14,48 @@
 <?import javafx.scene.layout.Region?>
 <?import javafx.scene.layout.VBox?>
 
+<?import org.controlsfx.control.tableview2.FilteredTableView?>
+<?import javafx.scene.layout.Pane?>
 <BorderPane xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.owlplug.core.controllers.PluginsController">
-   <center>
+    <top>
+        <HBox alignment="CENTER_LEFT" prefHeight="40.0" spacing="5">
+            <Button fx:id="newLinkButton" alignment="BASELINE_CENTER" text="New link">
+                <graphic>
+                    <Region prefHeight="15.0" prefWidth="18.0"
+                            style="-fx-background-color: white; -fx-shape: &quot; m 19.847458,19.194915 4.152543,-2.943321 -4.152543,-2.870237 v 1.263203 c 0,0 -4.824423,1.072601 -5.813559,5.380864 2.106169,-2.255246 5.813559,-2.166797 5.813559,-2.166797 z M 19.5,13 c -13,7.333333 -6.5,3.666667 0,0 z m -4.68,9 H 0 V 2 h 7 c 1.695,1.942 2.371,3 4,3 h 13 v 7.82 C 22.831,11.696 21.246,11 19.5,11 15.916,11 13,13.916 13,17.5 c 0,1.747 0.695,3.331 1.82,4.5 z&quot;;"/>
+                </graphic>
+            </Button>
+            <TextField fx:id="searchTextField" prefHeight="27.0" prefWidth="170.0" promptText="Search"
+                       HBox.hgrow="ALWAYS"/>
+            <Pane HBox.hgrow="ALWAYS" />
+            <Button fx:id="exportButton" text="Export">
+                <graphic>
+                    <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
+                        <image>
+                            <Image url="@../icons/connect-white-16.png"/>
+                        </image>
+                    </ImageView>
+                </graphic>
+            </Button>
+            <Button fx:id="syncButton" text="Sync">
+                <graphic>
+                    <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
+                        <image>
+                            <Image url="@../icons/refresh-white-16.png"/>
+                        </image>
+                    </ImageView>
+                </graphic>
+            </Button>
+        </HBox>
+    </top>
+    <center>
       <BorderPane BorderPane.alignment="CENTER">
       	<center>
-      		<TreeView fx:id="pluginTreeView" prefHeight="215.0" prefWidth="521.0" HBox.hgrow="SOMETIMES" VBox.vgrow="ALWAYS">
+            <VBox fx:id="pluginsContainer">
+                <FilteredTableView fx:id="pluginTableView">
 
-            </TreeView>
+                </FilteredTableView>
+            </VBox>
       	</center>
       	<left>
             <VBox styleClass="small-tab-pane-container">
@@ -29,53 +64,20 @@
                         <tabs>
                             <Tab fx:id="treeTabAll" text="All" closable="false" />
                             <Tab fx:id="treeTabDirectories" text="Directories" closable="false" />
+                            <Tab fx:id="tableTab" text="Table" closable="false" />
                         </tabs>
                     </TabPane>
                 </children>
             </VBox>
       	</left>
-      	<top>
-            <HBox alignment="CENTER_LEFT" prefHeight="40.0" spacing="5">
-               <children>
-                  <Button fx:id="newLinkButton" alignment="BASELINE_CENTER" text="New link">
-                     <graphic>
-                        <Region prefHeight="15.0" prefWidth="18.0" style="-fx-background-color: white; -fx-shape: &quot; m 19.847458,19.194915 4.152543,-2.943321 -4.152543,-2.870237 v 1.263203 c 0,0 -4.824423,1.072601 -5.813559,5.380864 2.106169,-2.255246 5.813559,-2.166797 5.813559,-2.166797 z M 19.5,13 c -13,7.333333 -6.5,3.666667 0,0 z m -4.68,9 H 0 V 2 h 7 c 1.695,1.942 2.371,3 4,3 h 13 v 7.82 C 22.831,11.696 21.246,11 19.5,11 15.916,11 13,13.916 13,17.5 c 0,1.747 0.695,3.331 1.82,4.5 z&quot;;" />
-                     </graphic>
-                  </Button>
-                <TextField fx:id="searchTextField" prefHeight="27.0" prefWidth="170.0" promptText="Search" HBox.hgrow="ALWAYS" />
-               </children>
-            </HBox>
-      	</top>
       </BorderPane>
    </center>
    <right>
-      <VBox prefHeight="200.0" prefWidth="100.0" BorderPane.alignment="CENTER">
+      <VBox fx:id="pluginInfoPane" prefHeight="200.0" prefWidth="100.0" BorderPane.alignment="CENTER">
          <children>
-            <HBox alignment="CENTER_RIGHT" minHeight="-Infinity" prefHeight="38.0" prefWidth="500.0" VBox.vgrow="NEVER" spacing="5.0">
-               <children>
-                   <Button fx:id="exportButton" text="Export">
-                       <graphic>
-                           <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
-                               <image>
-                                   <Image url="@../icons/connect-white-16.png" />
-                               </image>
-                           </ImageView>
-                       </graphic>
-                   </Button>
-                  <Button fx:id="syncButton" text="Sync">
-                     <graphic>
-                        <ImageView fitHeight="16.0" fitWidth="16.0" pickOnBounds="true" preserveRatio="true">
-                           <image>
-                              <Image url="@../icons/refresh-white-16.png" />
-                           </image>
-                        </ImageView>
-                     </graphic>
-                  </Button>
-               </children>
-            </HBox>
             <AnchorPane VBox.vgrow="ALWAYS">
                <children>
-            				<fx:include source="NodeInfoView.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
+                   <fx:include source="NodeInfoView.fxml" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
                </children>
             </AnchorPane>
          </children>
